feat: update board layout and styles
- Refactored board-group.component.html to improve loading and error
  states.
- Enhanced board layout with new LayoutComponent for better structure.
- Updated action-menu to append menu to body for better UI handling.
- Adjusted column styles for improved visual consistency.
- Removed unused dark mode toggle functionality from board-group.
- Cleaned up SCSS files for better readability and maintainability.
