+++ board/src/app/app.routes.ts
@@ -1,4 +1,11 @@
 import { Routes } from '@angular/router';
 import { BoardGroupComponent } from './components/board';
+import { LayoutComponent } from './components/layout/layout/layout.component';
 
-export const routes: Routes = [{ path: '', component: BoardGroupComponent }];
+export const routes: Routes = [
+  {
+    path: '',
+    component: LayoutComponent,
+    children: [{ path: '', component: BoardGroupComponent }],
+  },
+];
