+++ board/src/app/components/board/action-menu/action-menu.component.html
@@ -1,4 +1,4 @@
-<p-menu #menu [model]="items" [popup]="true" />
+<p-menu #menu [model]="items" appendTo="body" [popup]="true" />
 <p-button
   (click)="menu.toggle($event)"
   icon="pi pi-ellipsis-v"
