+++ board/src/app/components/board/board-group/board-group.component.scss
@@ -1,3 +1,9 @@
+:host {
+  //display: flex;
+  //flex-direction: column;
+  //flex-grow: 1;
+}
+
 .cdk-drop-list-dragging {
   transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
