+++ board/src/app/components/board/board-group/board-group.component.ts
@@ -3,7 +3,6 @@ import {
   Component,
   inject,
   OnInit,
-  signal,
 } from '@angular/core';
 import { BoardColumnComponent } from '../board-column';
 import { ColumnService } from '../services';
@@ -14,11 +13,10 @@ import {
 } from '@angular/cdk/drag-drop';
 import { Column } from '../models';
 import { ProgressSpinner } from 'primeng/progressspinner';
-import { Button } from 'primeng/button';
 
 @Component({
   selector: 'board-group',
-  imports: [DragDropModule, BoardColumnComponent, ProgressSpinner, Button],
+  imports: [DragDropModule, BoardColumnComponent, ProgressSpinner],
   templateUrl: './board-group.component.html',
   styleUrl: './board-group.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -27,7 +25,6 @@ export class BoardGroupComponent implements OnInit {
   columnService: ColumnService = inject(ColumnService);
 
   columns$ = this.columnService.columns;
-  iconClass = signal<string>('pi pi-moon');
 
   ngOnInit() {
     this.columnService.getColumns();
@@ -39,15 +36,5 @@ export class BoardGroupComponent implements OnInit {
       $event.previousIndex,
       $event.currentIndex,
     );
-  }
-
-  toggleDarkMode() {
-    const element = document.querySelector('html');
-    if (element !== null) {
-      element.classList.toggle('my-app-dark');
-      this.iconClass.set(
-        element.classList.contains('my-app-dark') ? 'pi pi-sun' : 'pi pi-moon',
-      );
-    }
   }
 }
