+++ board/src/app/components/layout/layout/layout.component.html
@@ -1,0 +1,6 @@
+<topbar class="w-full fixed top-0"></topbar>
+<div class="flex flex-col justify-between px-8">
+  <div class="min-h-screen pt-16">
+    <router-outlet></router-outlet>
+  </div>
+</div>
