+++ board/src/app/components/layout/layout/layout.component.ts
@@ -1,0 +1,12 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { TopbarComponent } from '../topbar/topbar.component';
+import { RouterOutlet } from '@angular/router';
+
+@Component({
+  selector: 'layout',
+  imports: [RouterOutlet, TopbarComponent],
+  templateUrl: './layout.component.html',
+  styleUrl: './layout.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class LayoutComponent {}
